# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paper Company",
    "version": "1.0.0",
    "description": "Witaj w repozytorium Systemu ERP dla Papier Company - zainspirowanego przez kultowy serial \"The Office\". Nasz system jest zaprojektowany, aby wspierać i optymalizować kluczowe procesy biznesowe w firmie zajmującej się dystrybucją papieru, od zarządzania zapasami po finanse i księgowość."
  },
  "paths": {
...
    "/admin": {
      "get": {
        "summary": "Pobiera listę wszystkich użytkowników.",
        "description": "Endpoint służący do pobierania listy wszystkich użytkowników.",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano listę użytkowników.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usersList": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera podczas pobierania listy użytkowników.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/{id}/{role}": {
      "put": {
        "summary": "Aktualizuje rolę użytkownika.",
        "description": "Endpoint służący do aktualizacji roli użytkownika na podstawie jego identyfikatora.",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role",
            "required": true,
            "description": "Nowa rola użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie zaktualizowano rolę użytkownika."
          },
          "500": {
            "description": "Błąd serwera podczas aktualizacji roli użytkownika.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/{id}": {
      "delete": {
        "summary": "Usuwa użytkownika.",
        "description": "Endpoint służący do usuwania użytkownika na podstawie jego identyfikatora.",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie usunięto użytkownika."
          },
          "500": {
            "description": "Błąd serwera podczas usuwania użytkownika.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/sales": {
      "get": {
        "summary": "Pobiera listę wszystkich zamowien nad ktorymi pracuje dany sprzedawca.",
        "description": "Endpoint służący do pobierania listy wszystkich sprzedazy.",
        "tags": [
          "Sales"
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano listę sprzedazy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ordersList": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera podczas pobierania listy użytkowników.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Dodaje nowe sprzedarze.",
        "description": "Jak wyzej.",
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imie": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "produkt": {
                    "type": "string"
                  },
                  "ilosc": {
                    "type": "string"
                  },
                  "miasto": {
                    "type": "string"
                  },
                  "ulica": {
                    "type": "string"
                  },
                  "nr_budynku": {
                    "type": "string"
                  },
                  "nr_mieszkania": {
                    "type": "string"
                  },
                  "kod": {
                    "type": "string"
                  },
                  "nazwa_firmy": {
                    "type": "string"
                  },
                  "salesId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano listę zamowien."
          },
          "400": {
            "description": "Błąd w żądaniu."
          },
          "401": {
            "description": "Nieautoryzowany dostęp."
          },
          "500": {
            "description": "Wystąpił błąd serwera."
          }
        }
      }
    },
    "/sales/{id}": {
      "delete": {
        "summary": "Usuwa zamowienia.",
        "description": "Endpoint służący do usuwania zamowienia na podstawie jego identyfikatora.",
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator zamowienia.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie usunięto zamowienie."
          },
          "500": {
            "description": "Błąd serwera podczas usuwania użytkownika.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}